V1.11.0 - Add timestamping to replay comments
Allow users to use timecodes in replay comments that allow for jumping directly to a timestamp in the player without a reload.
This will require a transition to a new API endpoint that handles processing our body text, parsing for timecodes to convert to links. Also supports a variety of other components.