Throttle viewerlist update frequency
Previously it was reasonable to allow the chat server to notify via socket when the user list had updated.
The chat server itself uses the same throttling function even.
However now there are 20 instances of the chat server, and throttling is done per instance.
This commit throttles the update frequency to once every 15 seconds, which should be reasonable.
This will reduce the amount of requests the client will have to manage.
The new throttle function is a handy general utility too.
It's overly documented in a giant comment above the function.

there's more documenation than code oh god I'm sorry.