v1.24.3 - Adds new popup dialog UI for ignore lists
Allows for users to see and unignore both users and channels
Uses a dialog UI displaying a data table of userames for each
Reuses existing functions to unignore
Creates new dialog that is a relative to it's parent.
Shows overlay and dialog only over it's parent, not entire interface.

Could be modularized for future usage with more dialogs