Streamer Channel Icons
Adds a animated streamer icon next to the streamer's name in chat when a streamer talks in their own channel.
Later on, this slot can be used to hold a badge for channel mods, and a custom channel badges for a few loyal / supporting viewers of each channel.
Adds fun css text animation classes.